Bug dish updated corrected

+++ src/controllers/DishesController.js
@@ -4,15 +4,27 @@ const AppError = require("../utils/AppError");
 
 class DishesController {
   async create(request, response) {
-    const { name, description, category, price, ingredients } = request.body
-    const image = request.file.filename
-    const user_id = request.user.id
-
+    const { name, description, category, price, ingredients } = request.body;
+    const image = request.file.filename;
+    const user_id = request.user.id;
+  
     const diskStorage = new DiskStorage();
     const filename = await diskStorage.saveFile(image);
-
-    const ingredientsArray = JSON.parse(ingredients || '[]');
-
+  
+    // Verifique se ingredients é uma string e faça o parsing, caso contrário, use-o diretamente
+    let ingredientsArray = [];
+    if (typeof ingredients === 'string') {
+      try {
+        ingredientsArray = JSON.parse(ingredients);
+      } catch (error) {
+        throw new AppError("Formato inválido para os ingredientes.", 400);
+      }
+    } else if (Array.isArray(ingredients)) {
+      ingredientsArray = ingredients;
+    } else {
+      throw new AppError("Ingredientes devem ser uma string JSON ou um array.", 400);
+    }
+  
     const [dish_id] = await knex("dishes").insert({
       name,
       description,
@@ -21,22 +33,20 @@ class DishesController {
       image: filename,
       created_by: user_id,
       updated_by: user_id,
-      
     });
-
+  
     const ingredientsInsert = ingredientsArray.map((name) => {
       return {
         dish_id,
         name,
         created_by: user_id,
-      }
-    })
-
+      };
+    });
+  
     await knex("ingredients").insert(ingredientsInsert);
-
+  
     return response.json();
   }
-
   async show(request, response) {
     const { id } = request.params;
 
@@ -63,13 +73,13 @@ class DishesController {
     const { id } = request.params;
     const { name, description, category, price, ingredients } = request.body;
     const imageFilename = request.file?.filename;
-
+  
     const dish = await knex("dishes").where({ id }).first();
-
+  
     if (!dish) {
       throw new AppError("Prato não encontrado.", 404);
     }
-
+  
     const dishUpdate = {
       name: name ?? dish.name,
       description: description ?? dish.description,
@@ -78,37 +88,56 @@ class DishesController {
       updated_by: request.user.id,
       updated_at: knex.fn.now(),
     };
-
+  
     if (imageFilename) {
       const diskStorage = new DiskStorage();
-
+  
       if (dish.image) {
         await diskStorage.deleteFile(dish.image);
       }
-
+  
       const filename = await diskStorage.saveFile(imageFilename);
       dishUpdate.image = filename;
     }
-
+  
     if (ingredients) {
+      // Verifique se 'ingredients' é uma string e tente convertê-la em um array
+      let ingredientsArray = [];
+      if (typeof ingredients === 'string') {
+        try {
+          ingredientsArray = JSON.parse(ingredients);
+        } catch (error) {
+          throw new AppError("Formato de ingredientes inválido.", 400);
+        }
+      } else if (Array.isArray(ingredients)) {
+        ingredientsArray = ingredients;
+      } else {
+        throw new AppError("Formato de ingredientes inválido.", 400);
+      }
+  
+      // Certifique-se de que 'ingredientsArray' é realmente um array antes de usar 'map'
+      if (!Array.isArray(ingredientsArray)) {
+        throw new AppError("Formato de ingredientes inválido.", 400);
+      }
+  
       await knex("ingredients").where({ dish_id: id }).delete();
-
-      const ingredientsInsert = ingredients.map((name) => {
+  
+      const ingredientsInsert = ingredientsArray.map((name) => {
         return {
           dish_id: id,
           name,
           created_by: dish.created_by,
         };
       });
-
+  
       await knex("ingredients").insert(ingredientsInsert);
     }
-
+  
     await knex("dishes").where({ id }).update(dishUpdate);
-
+  
     return response.json();
   }
-
+  
   async index(request, response) {
     const { search } = request.query;
 
@@ -165,53 +194,5 @@ class DishesController {
 
     return response.json(dishesWithIngredients);
   }
-
-    
-      
-   /*  async index(request, response){
-        const { title, tags } = request.query
-        const user_id = request.user.id
-
-        let dishes
-
-        if( tags ){
-            const filterTags = tags.split(',').map((tag) => tag.trim())
-
-            dishes = await knex("tags")
-            .select([
-                "dishes.id",
-                "dishes.title",
-                "dishes.user_id"
-            ])
-            .where("dishes.user_id", user_id)
-            .whereLike("dishes.title", `%${title}%`)
-            .whereIn("name", filterTags)
-            .innerJoin("dishes", "dishes.id", "tags.dish_id")
-            .orderBy("dishes.title")
-
-        }else{
-            dishes = await knex("dishes")       
-            .where({ user_id })
-            .whereLike("title", `%${title}%`)       
-            .orderBy("title");
-        }        
-    
-        const userTags = await knex("tags").where({ user_id });
-
-        const dishesWithTags = dishes.map((dish) => {
-        const dishTags = userTags.filter((tag) => tag.dish_id === dish.id);
-            
-            return {
-                ...dish,
-                tags: dishTags,
-            }
-        
-        })
-
-        return response.json({ dishesWithTags })
-    } */
-
-
-
-      }
+}
       module.exports = DishesController;
